+++ public/assignment/index.html
@@ -8,10 +8,11 @@
           integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u"
           crossorigin="anonymous">
     <link href="css/styles.css" rel="stylesheet"/>
+    <script src="../vendor/angular-route.js"></script>
+    <script src="https://ajax.googleapis.com/ajax/libs/angularjs/1.6.4/angular.min.js"></script>
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
     <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
     <script src="https://ajax.googleapis.com/ajax/libs/angularjs/1.6.4/angular.min.js"></script>
-    <script src="../vendor/angular-route.js"></script>
     <script src="js/app.js"></script>
     <script src="js/config.js"></script>
     <script type="text/javascript" src="views/user/controllers/login.controller.client.js"></script>
@@ -30,6 +31,10 @@
     <script type="text/javascript" src="services/page.service.client.js"></script>
     <script type="text/javascript" src="services/website.service.client.js"></script>
     <script type="text/javascript" src="services/widget.service.client.js"></script>
+    <script type="text/javascript" src="../../assignment/services/user.service.server.js"></script>
+    <script type="text/javascript" src="../../assignment/services/website.service.server.js"></script>
+    <script type="text/javascript" src="../../assignment/services/page.service.server.js"></script>
+    <script type="text/javascript" src="../../assignment/services/widget.service.server.js"></script>
 
 
 </head>
