+++ public/assignment/views/website/controllers/website-edit.controller.client.js
@@ -17,7 +17,12 @@
         vm.backToList = backToList;
 
         function init() {
-            vm.website = WebsiteService.findWebsiteById(wid);
+
+            WebsiteService
+                .findWebsiteById(wid)
+                .then(function (response) {
+                    vm.website = response.data;
+                });
         }
 
         init();
