+++ server.js
@@ -5,24 +5,11 @@ var bodyParser = require('body-parser');
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({extended: true}));
 
-app.set('view engine', 'ejs');
-// require("./utilities/filelist");
-
 app.use(express.static(__dirname + '/public'));
 
 require("./test/app");
 
+require("./assignment/app")(app);
 
-var express = require('express');
-var app = express();
-
-// install, load, and configure body parser module
-var bodyParser = require('body-parser');
-app.use(bodyParser.urlencoded({ extended: true }));
-app.use(bodyParser.json());
-
-require("./public/assignment/js/app")(app);
-
-var port = process.env.PORT || 3000;
-
+port = process.env.PORT || 3000;
 app.listen(port);
